requirements.txt
fastapi
uvicorn[standard]
celery[redis]
sqlalchemy
psycopg2-binary
pydantic-settings
python-multipart
aiofiles
redis[hiredis]
pyannote.audio
speechbrain
torch
torchaudio
requests
numpy
scipy
